Inpainting now working with sequencer
Added model presets. Viewport capture is now handled by a SceneCapture2D instead of directly from the viewport. This unfortunately has led to the default images being too dark due to the automatic exposure not being captured. The temporary workaround is to make sure that the global post-processing exposure settings are set to manual and a bias of 15.